requirements.txt
drawSvg==1.8.3
Pillow==9.2.0
numpy==1.23.2